create with product reusable component added
