+++ components/pages/stock/Damaged/index.jsx
@@ -3,13 +3,15 @@ import { sidebarDatas } from "@/layout/utils/sidebarDatas";
 import BreadCrumb from "@/components/reusable/Breadcrumb";
 import { damagedStock } from "@/contents/bengali";
 import Head from "next/head";
-import ProductList from "./ProductList";
-import SelectedProduct from "./SelectedProduct";
+import { useDamagedStockData } from "./useDamagedStock";
+import CreateWithProduct from "@/components/reusable/CreateWithProduct";
 
 const DamagedStock = () => {
   const router = useRouter();
   const pathname = "/" + router?.pathname.split("/")[1];
   const previousPages = sidebarDatas()?.filter((item) => item?.url === pathname);
+
+  const { mutate, isLoading, isError, error } = useDamagedStockData();
   return (
     <>
       <Head>
@@ -19,10 +21,7 @@ const DamagedStock = () => {
         previousPages={previousPages}
         currentPage={damagedStock?.pageTitle}
       />
-      <div className="mt-5 px-4 py-2 grid grid-cols-1 md:grid-cols-2 gap-2">
-        <ProductList />
-        <SelectedProduct />
-      </div>
+      <CreateWithProduct incQty={false} mutate={mutate} isLoading={isLoading}/>
     </>
   );
 };
