+++ components/pages/stock/New/index.jsx
@@ -3,13 +3,15 @@ import { sidebarDatas } from "@/layout/utils/sidebarDatas";
 import BreadCrumb from "@/components/reusable/Breadcrumb";
 import { newStock } from "@/contents/bengali";
 import Head from "next/head";
-import ProductList from "./ProductList";
-import SelectedProduct from "./SelectedProduct";
+import CreateWithProduct from "@/components/reusable/CreateWithProduct";
+import { useNewStockData } from "./useNewStock";
 
 const NewStock = () => {
   const router = useRouter();
   const pathname = "/" + router?.pathname.split("/")[1];
   const previousPages = sidebarDatas()?.filter((item) => item?.url === pathname);
+
+  const { mutate, isLoading, isError, error } = useNewStockData();
   return (
     <>
       <Head>
@@ -19,10 +21,7 @@ const NewStock = () => {
         previousPages={previousPages}
         currentPage={newStock?.pageTitle}
       />
-      <div className="mt-5 px-4 py-2 grid grid-cols-1 md:grid-cols-2 gap-2">
-        <ProductList />
-        <SelectedProduct />
-      </div>
+      <CreateWithProduct mutate={mutate} isLoading={isLoading} />
     </>
   );
 };
